% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factor.R
\name{factor}
\alias{factor}
\title{Estimate number of factors}
\usage{
factor(X, rmax)
}
\arguments{
\item{X}{a given matrix}

\item{rmax}{the largest number of factor}
}
\value{
the number of factor
}
\description{
return the number of factors
}
