% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gibbs}
\alias{Gibbs}
\title{Gibbs}
\usage{
Gibbs(n_samples, n, a, b)
}
\arguments{
\item{n_samples}{samples}

\item{n}{a int value}

\item{a}{a double value1}

\item{b}{a double value2}
}
\value{
List
}
\description{
Gibbs sampler
}
