% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LossFunction}
\alias{LossFunction}
\title{Loss Function}
\usage{
LossFunction(ytrue, ypred, alpha)
}
\arguments{
\item{ytrue}{true vector}

\item{ypred}{predicted vector}

\item{alpha}{weight}
}
\value{
loss
}
\description{
calculate loss funtion
}
