\name{SA24204149-package}
\alias{SA24204149-package}
\alias{SA24204149}
\docType{package}
\title{
  my package for statcomp
}
\description{
 Functions used to estimate number of factors and calculate loss funtion.
}
\details{
  estimate number of factors and calculate loss funtion.
}
\author{
Huang Yue, hy020917@mail.ustc.edu.cn.

Maintainer: Huang Yue <hy020917@mail.ustc.edu.cn>
}
\keyword{ package }
